Bhargava-PayrollApplication/requirements.txt


fastapi
uvicorn
python-jose
passlib[argon2]
argon2-cffi
pydantic
django
djangorestframework
mysqlclient
